Rename Workflow package to automations
